+++ src/database/migrations/2015_05_18_034011_add_active_to_pricelists_table.php
@@ -1,0 +1,40 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddActiveToPricelistsTable extends Migration {
+
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+        // Add column "active" if NOT exist
+        if (!Schema::hasColumn('pricelists', 'active')) {
+            Schema::table('pricelists', function(Blueprint $table)
+            {
+                $table->boolean('active')->default(true);
+            });
+        }
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+        // Drop column "active" if exists
+        if (Schema::hasColumn('pricelists', 'active')) {
+            Schema::table('pricelists', function(Blueprint $table)
+            {
+                $table->dropColumn('active');
+            });
+        }
+	}
+
+}